refactor: Drop enhancer API in favor of separate middlewares+reducer
BREAKING CHANGE: makeMidiEnhancer has been removed. It became increasingly difficult to reason about its correctness with respect to other enhancers and particularly user-specified middleware. The new API is a little more verbose but allows the user to place MIDI I/O explicitly in the middleware chain as well as explicitly mount the required reducer into their state tree.
Tests, documentation and the bare-input example are up-to-date.